+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Methods:
 
 * `kala.text.StringSlice::split(String)`
 * `kala.text.StringSlice::split(String, int)`
+* `kala.collection.mutable.MutableSeq::create(int)`
 * `kala.collection.mutable.MutableSeq::setAll(E...)`
 
 ### Removed